fix: fixed the required changes
